sql/rowexec: move code used only for testing to test file
Release note: None
